bridge/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
python-docx==1.1.2
